backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary
django-ninja==1.4.3
pydantic==2.11.9
djangorestframework >=3.15
djangorestframework-simplejwt >=5.5.1,<6

# Dev / testing
pytest
pytest-django
httpx
